requirements.txt
# MCP Framework Dependencies
# Install with: pip install -r requirements.txt

# Core Flask
Flask>=3.0.0
flask-cors>=4.0.0

# Database
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0

# Authentication
PyJWT>=2.8.0

# HTTP Requests
requests>=2.31.0

# Google Analytics (optional)
google-analytics-data>=0.18.0
google-auth>=2.23.0

# Environment variables
python-dotenv>=1.0.0

# Production server
gunicorn>=21.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.1.0
